Updated Codes

+++ Mabolo-Dormitory-System/roomTab.cs
@@ -15,26 +15,32 @@ namespace Mabolo_Dormitory_System
 {
     public partial class roomTab : UserControl
     {
-        public int roomNum = 0;
-        public DatabaseManager db;
+        private int roomNum = 0;
+        private DatabaseManager db;
         private List<User> users;
-      
+        
         public roomTab()
         {
             db = new DatabaseManager();
             InitializeComponent();
+
+            // Set up Room Choose ComboBox
             for (int i = 1; i < 10; i++)
                 roomChooseCB.Items.Add("Room " + i);
             roomChooseCB.Text = "1";
-           
+            
+            // Set up Room User Default View 
             users = db.GetUsersInRoom(roomNum);
-            SetUpTable(bigBrodView, users);
+            SetUpTable(roomUserView, users);
 
         }
         private void SetUpTable(GunaDataGridView view, List<User> users)
         {
+            // Reset Table
             view.Rows.Clear();
             view.Columns.Clear();
+
+            // Set up Table
             DataGridViewCheckBoxColumn checkBoxColumn = new DataGridViewCheckBoxColumn();
             checkBoxColumn.Name = "Column1";
             checkBoxColumn.HeaderText = "#";
@@ -46,28 +52,29 @@ namespace Mabolo_Dormitory_System
             view.Columns.Add("Email", "Email");
             view.Columns.Add("PhoneNumber", "Phone Number");
             view.Columns.Add("UserType", "User Type");
+
+            // Set up Table Style
             view.Columns["Column1"].Width = 50;
             view.DefaultCellStyle.Font = new Font("Century Gothic", 12);
             view.DefaultCellStyle.ForeColor = Color.Black;
             view.AllowUserToResizeRows = false;
-
             view.ColumnHeadersDefaultCellStyle.SelectionForeColor = Color.White;
             view.ColumnHeadersDefaultCellStyle.SelectionBackColor = Color.FromArgb(40, 167, 69);
             view.ColumnHeadersDefaultCellStyle.Font = new Font("Century Gothic", 14, FontStyle.Bold);
-
             view.DefaultCellStyle.Alignment = DataGridViewContentAlignment.MiddleCenter;
             view.ColumnHeadersDefaultCellStyle.Alignment = DataGridViewContentAlignment.MiddleCenter;
             view.RowTemplate.Height = 40;
 
+            // Add ComboBox Column
             DataGridViewComboBoxColumn comboBoxColumn = new DataGridViewComboBoxColumn();
             comboBoxColumn.Items.AddRange("Update", "View");
             comboBoxColumn.Name = "Action";
             comboBoxColumn.HeaderText = "Action";
             comboBoxColumn.ValueType = typeof(String);
             view.Columns.Add(comboBoxColumn);
-
             view.EditingControlShowing += new DataGridViewEditingControlShowingEventHandler(dataGridView1_EditingControlShowing);
-
+            
+            // Add Data to Table
             foreach (DataGridViewColumn column in view.Columns)
             {
                 column.SortMode = DataGridViewColumnSortMode.NotSortable;
@@ -78,6 +85,7 @@ namespace Mabolo_Dormitory_System
                 view.Rows.Add(false, user.UserId, user.FirstName, user.LastName, user.Birthday.ToString("yyyy/MM/dd"), user.Email, user.PhoneNumber, user.UserStatus);
             }
         }
+
         private void dataGridView1_EditingControlShowing(object sender, DataGridViewEditingControlShowingEventArgs e)
         {
             ComboBox combo = e.Control as ComboBox;
@@ -91,17 +99,15 @@ namespace Mabolo_Dormitory_System
         private void ComboBox_SelectedIndexChanged(object sender, EventArgs e)
         {
             ComboBox cb = (ComboBox)sender;
-          // MessageBox.Show(sender.ToString());
             if (cb.SelectedItem != null)
             {
-                int i = bigBrodView.CurrentRow.Index;
+                int i = roomUserView.CurrentRow.Index;
                 if (cb.SelectedItem.ToString() == "Update")
                 {
                     UpdateForm update = new UpdateForm();
                     SetFormLocation(update);
                     update.SetInformation(users, i);
                     update.Show();
-                   // cb.SelectedIndexChanged -= new EventHandler(ComboBox_SelectedIndexChanged);
                 }
                 else if (cb.SelectedItem.ToString() == "View")
                 {
@@ -109,39 +115,26 @@ namespace Mabolo_Dormitory_System
                     SetFormLocation(viewForm);
                     viewForm.SetInformation(users, i);
                     viewForm.Show();
-                  //  cb.SelectedIndexChanged -= new EventHandler(ComboBox_SelectedIndexChanged);
                 }
-               
             }
         }
+
         private void CreateTable(GunaDataGridView view, List<User> users)
         {
             view.Rows.Clear();
-            
             foreach (User user in users)
-            {
-                
+            { 
                 view.Rows.Add(false, user.UserId, user.FirstName, user.LastName, user.Birthday.ToString("yyyy/MM/dd"), user.Email, user.PhoneNumber, user.UserType);
-                
-            }
-            
-        }
-        private void roomChooseCB_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            String text = roomChooseCB.SelectedItem.ToString();
-            text = text.Split()[1];
-            roomNum = Convert.ToInt32(text);
-            //MessageBox.Show("Room " + roomNum);
-            CreateTable(bigBrodView, db.GetUsersInRoom(roomNum));
-           
+            }  
         }
 
-        private void add2_Click(object sender, EventArgs e)
+        private void roomChooseCB_SelectedIndexChanged(object sender, EventArgs e)
         {
-            AddRoomAllocation add = new AddRoomAllocation(roomNum, "Regular Dormer");
-            SetFormLocation(add);
-            add.Show();
+            String text = roomChooseCB.SelectedItem.ToString().Split()[1];
+            roomNum = Convert.ToInt32(text);
+            CreateTable(roomUserView, db.GetUsersInRoom(roomNum)); 
         }
+
         private void SetFormLocation(Form form)
         {
             form.StartPosition = FormStartPosition.Manual;
@@ -158,39 +151,53 @@ namespace Mabolo_Dormitory_System
             
         }
 
-
         private void refreshBut_Click(object sender, EventArgs e)
         {
-            CreateTable(bigBrodView, db.GetUsersInRoom(roomNum));
+            CreateTable(roomUserView, db.GetUsersInRoom(roomNum));
         }
 
         private void delBut_Click(object sender, EventArgs e)
         {
-            bool isChecked = false;
-            for (int i = bigBrodView.Rows.Count - 1; i >= 0; i--)
+            MessageBoxButtons messageBoxButtons = MessageBoxButtons.YesNo;
+            bool isChecked = false;    
+            foreach (DataGridViewRow row in roomUserView.Rows)
             {
-                DataGridViewRow row = bigBrodView.Rows[i];
                 if (row.Cells["Column1"].Value != null && Convert.ToBoolean(row.Cells["Column1"].Value))
                 {
-                    isChecked = true;
-                    db.DeleteUser(row.Cells["UserId"].Value.ToString());
-                    bigBrodView.Rows.RemoveAt(i);
+                    DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete the selected rows?\nThis action cannot be undone!", "Delete", messageBoxButtons);
+                    if (dialogResult == DialogResult.Yes)
+                    {
+                        isChecked = true;
+
+                        db.DeleteUser(row.Cells["UserId"].Value.ToString());
+                        roomUserView.Rows.Remove(row);
+                    }
+                    else
+                    {
+                        MessageBox.Show("Selected row(s) not deleted");
+                        break;
+                    }
                 }
             }
-
             if (!isChecked)
-            {
                 MessageBox.Show("Please select a row(s) to delete");
-            }
+            
             else
-            {
                 MessageBox.Show("Selected row(s) deleted successfully");
-            }
+            
         }
 
-        private void bigBrodView_CellContentClick(object sender, DataGridViewCellEventArgs e)
+        private void roomUserView_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
-
+            var senderGrid = (DataGridView)sender;
+            if (senderGrid.Columns["Column1"] is DataGridViewCheckBoxColumn && e.RowIndex > -1)
+            {
+                if(!Convert.ToBoolean(senderGrid.Rows[e.RowIndex].Cells["Column1"].Value))
+                    senderGrid.Rows[e.RowIndex].Cells["Column1"].Value = true;
+   
+                else
+                    senderGrid.Rows[e.RowIndex].Cells["Column1"].Value = false;
+            }
         }
 
         private void editButton_Click(object sender, EventArgs e)
